moved is_shared_ptr traits to Utility.hpp
